requirements.txt
altair==5.5.0
attrs==25.1.0
Authlib==1.5.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
# Editable install with no version control (ceas==0.0.1)
-e /.
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
duckdb==1.2.1
entrypoints==0.4
et_xmlfile==2.0.0
Faker==37.0.0
favicon==0.7.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.1
gspread-pandas==3.3.0
htbuilder==0.9.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.3.1
Markdown==3.7
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.30.0
numpy==2.2.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
prometheus_client==0.21.1
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
pymdown-extensions==10.14.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
six==1.17.0
smmap==5.0.2
soupsieve==2.6
sql_metadata==2.15.0
sqlparse==0.5.3
st-annotated-text==4.0.2
st-gsheets-connection==0.1.0
st-theme==1.2.3
streamlit==1.43.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
